Big code commit, working for last couple days on this
Wanted to get development server up and running before committing anything, for that matter, kinda forgot to do a git add till now. Anyways, I went with node.js (like we specified) and created a bunch of routes; first for diagnosis and then for symptom. created a search endpoint as well, and used zod (asked chatgpt for a framework and this is what it said was the best) to make sure the search param makes sense. Then I used zod to verify all the current and future endpoints.