starting on registration/login
Here is the beginning of the authentication part of the backend. All authentication files start with 'auth'. The schema checking is from zod in 'usermodel' (obv to make sure that we don't insert some random crap into our database). I also added a file to utility/ for actually connecting to the psql database for adding users to it.