Fixed routing errors, login/phrase passing to backend. Fuzzy search.
Fixed up from errors in routing (by me), and then fixed the login passing used to commit users to the nurses_doctors database. Also implemented fuzzy search for symptoms, I still need to add 'selection' to the search so the symptoms can be commited to the db from the search rather than adding it manually.